WIP: Support objerve instances as properties
Now that we have prefix listening, we can do this clever meta twist.

Edge case:  What if what we got assigned as a property isn't actually
the root objerve object, but some sub-property of it?  Perhaps we have
to maintain a WeakMap of the root object of every object we see, so we
can check whether they're part of a proxy.  This is also why we need
prefix-listening specifically; just root listening won't really do.